Rusty Skylines Alpha v1.6.13a March 15th 23:09 NOT PI DAY!!
[Major]
Made entire prop rendering system and already render trees. (Not spawned yet)
It is instanced, of course and procedural, of course!
You can place trees using the new 'Props' cursor mode.

[Minor]
Changed how the Cursor mode stores data, like prop name or road specs.

Renamed more structs to simpler names. Like 'CarSubsystem' to 'Cars' and 'RoadSubsystem' to 'Roads' and 'AstronomyState' to just 'Astronomy'.
Uselessly long names are horrible to code and look at.

You can fetch the state of a setting using a variable in text like normal.
If the variable name you typed matches the setting name, you get the value! No underscores exist in the settings btw

Upgraded the CarChunkDistance and WorldPos::dist2() to 64 bit values.

***
Not pi day, but 15 day. half month day i guess?
random-ass day but cool day, prop day i shall call it! where is my CAPS HUH HOW DO I TURN THIS OFF?? OK IF I CAN'T TURN IT OFF;
THEN I WILL SCREAM OBSCENITIES: number 1: oh wait why did caps TURN OFF HOW IRONIC  THANKS FOR STICKING FOR MY PERFORMANCE; BYE!

+++ src/helpers/positions.rs
@@ -42,10 +42,10 @@ impl ChunkCoord {
         }
     }
     #[inline]
-    pub fn dist2(&self, other: &ChunkCoord) -> u32 {
+    pub fn dist2(&self, other: &ChunkCoord) -> u64 {
         let dx = self.x as i64 - other.x as i64;
         let dz = self.z as i64 - other.z as i64;
-        (dx * dx + dz * dz) as u32
+        (dx * dx + dz * dz) as u64
     }
 }
 impl LocalPos {
